pyproject.toml
[tool.poetry]
name = "python-web-template"
version = "0.1.0"
description = ""
authors = ["pray <prayfff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.2.1"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.1"
loguru = "^0.7.3"
pyinstrument = "^5.0.0"
fastapi = "^0.115.6"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
